introduce new internal DisqusConfig service
