+++ src/Helper/DisqusHelper.php
@@ -12,20 +12,21 @@
 namespace Knp\Bundle\DisqusBundle\Helper;
 
 use Knp\Bundle\DisqusBundle\Client\DisqusClientInterface;
+use Knp\Bundle\DisqusBundle\Model\DisqusConfig;
 use Twig\Environment;
 use Twig\Extension\RuntimeExtensionInterface;
 
 class DisqusHelper implements RuntimeExtensionInterface
 {
     private $twig;
+    private $config;
     private $disqus;
-    private $debug;
 
-    public function __construct(Environment $twig, DisqusClientInterface $disqus, bool $debug)
+    public function __construct(Environment $twig, DisqusConfig $config, DisqusClientInterface $disqus)
     {
         $this->twig = $twig;
+        $this->config = $config;
         $this->disqus = $disqus;
-        $this->debug = $debug;
     }
 
     public function render(string $shortname, array $parameters = [], $template = '@KnpDisqus/list.html.twig')
@@ -33,19 +34,19 @@ class DisqusHelper implements RuntimeExtensionInterface
         try {
             $content = $this->disqus->fetch($shortname, $parameters);
         } catch (\Exception $e) {
-            if ($this->debug) {
+            if ($this->config->isDebug()) {
                 $error = $e->getMessage();
             } else {
                 $error = 'Oops! Seems there are problem with access to disqus.com. Please refresh the page in a few minutes.';
             }
         }
 
-        $sso = $this->disqus->getSsoParameters($parameters);
+        $sso = $this->config->getSsoParameters($parameters);
 
         $parameters['shortname'] = $shortname;
         $parameters['error'] = $error ?? null;
         $parameters['content'] = $content ?? [];
-        $parameters = $parameters + $this->disqus->getParameters();
+        $parameters = $parameters + $this->config->getParameters();
         $parameters['sso'] = $sso;
 
         return $this->twig->render($template, $parameters);
