+++ src/Resources/config/services.xml
@@ -5,18 +5,21 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="knp_disqus.disqus_client" class="Knp\Bundle\DisqusBundle\Client\DisqusClient">
+        <service id="knp_disqus.disqus_config" class="Knp\Bundle\DisqusBundle\Model\DisqusConfig">
             <argument/> <!-- $apiKey -->
             <argument/> <!-- $secretKey -->
             <argument>%kernel.debug%</argument>
+        </service>
+        <service id="knp_disqus.disqus_client" class="Knp\Bundle\DisqusBundle\Client\DisqusClient">
+            <argument type="service" id="knp_disqus.disqus_config"/>
         </service>
 
         <service id="Knp\Bundle\DisqusBundle\Client\DisqusClientInterface" alias="knp_disqus.disqus_client" />
 
         <service id="knp_disqus.helper" class="Knp\Bundle\DisqusBundle\Helper\DisqusHelper">
             <argument type="service" id="twig"/>
+            <argument type="service" id="knp_disqus.disqus_config"/>
             <argument type="service" id="knp_disqus.disqus_client"/>
-            <argument>%kernel.environment%</argument>
         </service>
     </services>
 </container>
